feat(antigravity): replace gemini runtime with antigravity
Gemini CLI is dead and gemini now redirects to Antigravity CLI (agy).
agy reads a plugin's rules from rules/ and ignores AGENTS.md, CLAUDE.md
and GEMINI.md at the repo root, and it has no session-start hook, so
rules/AGENTS.md is now the only channel through which the trigger
table reaches the model.
