+++ tests/build.test.mjs
@@ -52,16 +52,16 @@ test('never removes a path absent from the previous manifest', () => {
 
 test('loads core plus every configured runtime, in that order', async () => {
   const { emitters } = await build(process.cwd())
-  assert.deepEqual(emitters.map(e => e.id), ['core', 'claude-code', 'codex', 'opencode', 'gemini', 'pi'])
+  assert.deepEqual(emitters.map(e => e.id), ['core', 'claude-code', 'codex', 'opencode', 'antigravity', 'pi'])
 })
 
 // W2: the merge → regions → manifest → plan sequencing used to live untested in
 // the CLI. These assert the assembly, not just its parts.
 test('assembles every generated path and reports no drift on a clean tree', async () => {
   const { files, write, remove } = await build(process.cwd())
-  for (const path of ['package.json', 'CLAUDE.md', 'AGENTS.md', 'GEMINI.md', 'README.md',
+  for (const path of ['package.json', 'CLAUDE.md', 'AGENTS.md', 'rules/AGENTS.md', 'README.md',
                       '.claude-plugin/plugin.json', '.codex-plugin/plugin.json', '.opencode/plugins/vibekit.js',
-                      '.opencode/INSTALL.md', 'gemini-extension.json', MANIFEST]) {
+                      '.opencode/INSTALL.md', 'plugin.json', MANIFEST]) {
     assert.ok(path in files, `${path} missing from assembled output`)
   }
   assert.deepEqual(write, [], 'clean tree must report no stale files')
